Skin_Cancer_Classification/requirements.txt
flask==3.1.0
tensorflow==2.19.0
numpy==2.1.3
Pillow==11.2.1
werkzeug==3.1.3
python==3.11.5
flask_sqlalchemy==3.1.1
flask_bcrypt==1.0.1



